requirements.txt
gym==0.25.1
matplotlib==3.5.3
numpy==1.23.1
pandas~=2.1.1
python-dotenv==0.20.0
requests~=2.31.0
scikit_learn==1.1.2
seaborn==0.11.2
ta==0.10.1
TA_Lib==0.4.24
tensorflow==2.9.1
protobuf==3.9.2
tensorflow_addons==0.17.1
typeguard==2.13.3
alpaca-py

keras~=2.9.0
scikit-learn~=1.1.2

yfinance~=0.2.33
